backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
mysqlclient==2.2.7
scikit-surprise==1.1.4
scikit-learn==1.4.0
pandas==2.2.3
numpy==1.26.4
flask-cors==5.0.1
python-dotenv==1.1.0
werkzeug==3.1.3
